Overhauled password code to be class-based; Created PasswordGenerator.cpp
Created a new class called AccountManager that will be used by both client and server
to manage password and key stuff for login/authentication.
Broke out functionality into many modularized static functions.
Created a test Client that simulates creating a new user and logging in.
Changed the name of create_account() to insert_account() in SQLConnector.
Made sure that all strings are passed to and from the functions as char *.
Other minor code and comment changes.