Login works! ...kind of
Got data to be passed between server and client for login and get salt functions.
Changed generate_key so there is no need to free returned strings,
it no longer returns iterations, and a double pointer is no longer used.

TODO: The client needs to be restarted after an account create before a
login will work... There must be some kind of corruption going on during
create account...
TODO: It says the user successfully logged on when they didn't sometimes...