Fixed bug causing login to mess up after a create
I created code to listen to the response packet from a createAccount()
server call.

I think the issue was that the server was sending back a response packet,
and the client wasn't listening for it. That messed things up, though
I had assumed that it would be harmless. I think the packet somehow floated
to other recvfrom() calls, or it simply messed up the next one.