Aligned coding style of Michael's Funcs/Comments.
In an effort to keep a uniform coding style among all of the files in
the project, I added the default header to some of the classes that
michael made and also aligned function names to a thisIsAFunction()
style instead of this_is_a_function().  I also changed Utils.cpp so that
it is simply a namespace instead of a class with static function.  I
think this is much more common for including utility functions.