Got msgpack working with new packet header and classes
Got a class being converted to msgpack, sent to server, and
deserialized back to a class.
Created a new packet header format for all packets.
Altered Client and Server to test out the new packet headers and msgpack.
The makefile now automatically downloads and installs msgpack-c 2.1.1.
Created helper functions to help with the new packet headers.
Altered the packet size.
Added msgpack-c to .gitignore.
Updated the readme to reflect msgpack dependence.

Fixed small bug in server where a == was being used on a char *
instead of a std::string.

TODO: Expand the packet header to have some kind of checksum like fletcher-16?