All packets between Server and TestClient now use msgpack!
Lots of updates:
*Everything uses msgpack! Which means sending variable-sized data is
now a piece of cake! (assuming they all fit in ~500 bytes or less).
*Replaced all old packet structs with new packet classes.
*Told msgpack which fields to serialize inside the packet classes.
*Replaced as many fixed char * strings with std::string as possible.
*Fixed up some broken logic in the TestClient.
*Made the ListCharacters packet have a variable number of characters
by putting in a vector of strings - and it works in msgpack.
*Identified areas that need work.

Hardcoded some test characters in the server to show msgpack at its finest.

TODO: I believe there is a way to wrap all the msgpack and sendto stuff
into convenient and clean wrapper functions.