TestClient now uses new packets namespace; Removed PTYPE::
Realized we don't need PTYPE:: specifier, so removed it.
Part of the reason I wanted to go to the packets:: namespace was so I
could namespace the PTYPE enum without using an enum class.
Using an enum class would mean a lot of type casting that I don't
want to do.