Server/TestClient: Got message broadcasting working!
TestClient now has a packet queue architecture that allows it to
receive servercommand packets while also waiting for other packets!
This allows the server to send a broadcast packet without messing
up the execution of clients.
Created the BroadcastToConnections() Server method.
Created an activeConnections() map so the server can broadcast to
all clients.
Created a WaitOnPacketQueue() helper function that blocks on the
packet queue and uses the mutex and condition_variable libraries.