Got a unity client talking back and forth with the server!
Created a thread for every packet being sent (in addition to the main
thread and the PacketListener thread), so that it could properly
wait on the packetQueue without freezing unity.

While unity no longer blocks, the problem is that something is buggy
with how I use the threads I think. Maybe there is contention with
Debug.Log or something.

Another issue is how do I communicate back to the main Unity thread
safely if Unity can't block?

Also, the client accepts all incoming packet requests, when it should
only accept packets from the server and packets that have a valid
session id.