Unity Client: Got messages and broadcasts to display in the UI!!
Also, the Unity Client doesn't block anymore when it sends a packet
to the server and the server doesn't respond.

I changed the architecture to be more of a 'state machine.'
The disadvantage with this method, and partly why I avoided it for so
long, is because you can't just sit and block for a custom response.
To get around this, though, we can set state variables like
"isConnectAknowledged" or something, and so we know what the return
packet should do.