Removed the characters table; Improved npc import;
Realized that the Character table in the schema was a bad idea -
the sql schema should NOT try to look like our class inheritance.
So I kept the Character subclass, but db only deals with npcs and players.
Made it so that the npc import will autogenerate ids (so it can be done many times).
Updated items table to have player and npc ids.
Created views view_items_player and view_items_npc.